Add a check for thrift_installed when trying to define method_mappings in ThriftConnection so that rawes can work without Thrift support.
When Thrift is not available, the import of Method fails, which means that the attributes used to define the method_mappings dictionary are not available.  Since attempting to instantiate a ThriftConnection object while thrift_installed is True will raise an exception, having this dictionary empty in those cases should be completely safe.